+++ Lenguajes/Dotnet/ApuntesNet/BackendLibraries/HostedService/OtroHostedService.cs
@@ -1,0 +1,17 @@
+﻿using Microsoft.Extensions.Hosting;
+
+namespace HostedService;
+
+public class OtroHostedService : IHostedService
+{
+    public Task StartAsync(CancellationToken cancellationToken)
+    {
+        Console.WriteLine("Otro hosted service");
+        return Task.CompletedTask;
+    }
+
+    public Task StopAsync(CancellationToken cancellationToken)
+    {
+        throw new NotImplementedException();
+    }
+}
