+++ Lenguajes/Dotnet/ApuntesNet/BackendLibraries/HostedService/Program.cs
@@ -1,4 +1,5 @@
 ﻿using DependencyInjection;
+using HostedService;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
@@ -7,9 +8,10 @@ var builder = Host.CreateDefaultBuilder(args);
 builder.ConfigureServices(services =>
 {
     services.AddSingleton<IServicioInyectado, ServicioInyectado>();
-    services.AddHostedService<HostedService.HostedService>();
+    services.AddHostedService<BackgroundServiceTask>();
+    services.AddHostedService<OtroHostedService>();
 });
 
 var app = builder.Build();
 
-await app.StartAsync();
+await app.RunAsync();
